chore: move Button component file to common folder
