--- src/components/Button/Button.types.ts
@@ -1,7 +1,0 @@
-export type ButtonTheme = 'primary' | 'kakao';
-export type ButtonSize = 'small' | 'medium' | 'large' | 'responsive';
-
-export interface StyledButtonProps {
-  theme: ButtonTheme;
-  size: ButtonSize;
-}
