+++ src/stories/common/Button.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
-import Button, { ButtonProps } from '@components/Button/Button';
+import Button, { ButtonProps } from '@components/common/Button';
 
 const meta: Meta<ButtonProps> = {
   title: 'common/Button',
